+++ Source/Defcon/GameObjects/Enemies/bouncer.cpp
@@ -103,7 +103,7 @@ void Defcon::IBouncer::Explode(CGameObjectCollection& debris)
 	float fBrightBase;
 	this->CreateFireblast(debris, fBrightBase);
 
-	size_t a, i;
+	int32 a, i;
 	const float off = SFRAND * 0.2f;
 	const float off2 = SFRAND * 0.2f;
